Add ability to turn off auto-update
