+++ src/helpers/get-debug-info.js
@@ -9,23 +9,11 @@ const { getAppUpdateConfigSync } = require('../auto-updater')
 
 const packageJson = require(path.join(appDir, 'package.json'))
 const productName = require('./product-name')
-let electronBuilderConfig = {}
-
-try {
-  electronBuilderConfig = require(path.join(appDir, 'electron-builder-config'))
-} catch (err) {}
 
 let lastCommit = { hash: '-', date: '-' }
-let appUpdateConfig = {}
 
-try {
-  appUpdateConfig = getAppUpdateConfigSync()
-} catch (err) {
-  console.debug(err)
+const appUpdateConfig = getAppUpdateConfigSync()
 
-  appUpdateConfig = electronBuilderConfig
-    ?.publish ?? {}
-}
 try {
   lastCommit = require(path.join(appDir, 'lastCommit.json'))
 } catch (err) {
